refactor: chat 모듈 파일 구조 및 코드 정리 (#71)
