+++ src/components/chat/list/ChatListPanel.tsx
@@ -1,6 +1,5 @@
-import { ChatHeader } from '@/components/chat/common'
-import { ChatList } from '@/components/chat/list/ChatList'
-import type { ChatRoomListItem } from '@/types/chat'
+import { ChatHeader, ChatList } from '@/components/chat'
+import type { ChatRoomListItem } from '@/types'
 import { useState } from 'react'
 
 interface ChatListPanelProps {
@@ -27,7 +26,7 @@ export function ChatListPanel({
     <div className="flex h-full flex-col">
       <ChatHeader title="채팅방" onClose={onClose} />
       {hasNoRooms ? (
-        <div className="text-custom-gray-600 flex flex-1 flex-col items-center justify-center gap-1 text-sm">
+        <div className="text-custom-gray-600 centralize flex-1 flex-col gap-1 text-sm">
           <p>현재 참여 중인 채팅방이 없습니다.</p>
           <p>스터디에 들어가면 채팅방이 생깁니다.</p>
         </div>
